Order Process Done
The order pages are done, add item(s) to the cart, it will first check
if there's available stock then add to the cart. Then when you place
order it will add the new order to the customer's account. For now it
only shows the order ID, order date and order status, but all the
related items in that order are related back to the order id, just need
to make another page to display order details..

All is good in the hood bitches..

John out